Move ContentDisplay to a components/commons
This is because it is a commonly used component, that shall be used by
Materials as well.
