+++ src/components/commons/ContentDisplay.tsx
